# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/file_shares:
    get:
      tags:
        - FileShares
      summary: List FileShares
      description: Get a paginated list of FileShares with optional filtering and sorting
      operationId: listFileShares
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileShareList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - FileShares
      summary: Create FileShare
      description: Create a new FileShare
      operationId: createFileShare
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileShareRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileShare'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    FileShare:
      type: object
      required:
        - id
        - file_id
        - owner_id
        - token
        - share_type
        - permission
        - shared_with
        - download_count
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        file_id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        token:
          type: string
          minLength: 32
          maxLength: 64
        share_type:
          type: string
        permission:
          type: string
        shared_with:
          type: array
        password_hash:
          nullable: true
          type: string
        max_downloads:
          nullable: true
          type: integer
          format: int32
        download_count:
          type: integer
          format: int32
        expires_at:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        revoked_at:
          nullable: true
          type: string
          format: date-time
        revoked_by:
          nullable: true
          type: string
          format: uuid
        message:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateFileShareRequest:
      type: object
      required:
        - file_id
        - owner_id
        - token
        - share_type
        - permission
        - shared_with
        - download_count
        - status
        - metadata
      properties:
        file_id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        token:
          type: string
          minLength: 32
          maxLength: 64
        share_type:
          type: string
        permission:
          type: string
        shared_with:
          type: array
        password_hash:
          nullable: true
          type: string
        max_downloads:
          nullable: true
          type: integer
          format: int32
        download_count:
          type: integer
          format: int32
        expires_at:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        revoked_at:
          nullable: true
          type: string
          format: date-time
        revoked_by:
          nullable: true
          type: string
          format: uuid
        message:
          nullable: true
          type: string
        metadata:
          type: object
...
    FileShareList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FileShare'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []